refactor: renamend asset to classic
